% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aarset.R
\docType{data}
\name{aarset}
\alias{aarset}
\title{Failure times of devices from Aarset (1987)}
\format{
A vector of failure times (in hours)
}
\usage{
data(aarset)
}
\description{
Data on failure times from Aarset (1987)
}
\examples{
data(aarset)
\donttest{hist(aarset, 
               xlim=c(0,100), breaks=10, xlab="t (hours)", ylab="Frequency",
               main="Failure times of devices from Aarset (1987)")}
}
\references{
Aarset ML. How to Identify a Bathtub Hazard Rate.
In: IEEE Transactions on Reliability, R-36(1), 106-108, 1987.
}
\keyword{datasets}
