% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{as.parvec}
\alias{as.parvec}
\title{Convert model list object to parameter vector (incomplete)}
\usage{
as.parvec(model.list)
}
\arguments{
\item{model.list}{Model specification}
}
\description{
Convert model list object to parameter vector (incomplete)
}
