% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{both_lambda_func_dfr_c}
\alias{both_lambda_func_dfr_c}
\title{Hazard rate and integrated hazard rate functions - DFR}
\usage{
both_lambda_func_dfr_c(tvec, lambda0, thetavec, wvec)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}
}
\description{
Hazard rate function generated by 
a set of discrete locations and weights and integrated
}
