% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{both_lambda_func_lcv_c}
\alias{both_lambda_func_lcv_c}
\title{Hazard and integrated hazard rate functions - LCV}
\usage{
both_lambda_func_lcv_c(tvec, lambda0, w0, thetavec, wvec, epsilon)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{scaling}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}

\item{Offset}{to be added to the log hazard rate derivatice}
}
\description{
Hazard rate function generated by 
a set of discrete locations and weights and integrated
}
