% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chzf_con_c}
\alias{chzf_con_c}
\title{Integrated Hazard rate function - CON}
\usage{
chzf_con_c(tvec, lambda0)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}
}
\description{
Integrated hazard rate function for constant hazard
}
