% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chzf_hcv_c}
\alias{chzf_hcv_c}
\title{Integrated Hazard rate function - HCV}
\usage{
chzf_hcv_c(tvec, lambda0, a, thetavec, wvec)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{a}{Change point}

\item{thetavec}{Locations}

\item{wvec}{Weights}
}
\description{
Integrated hazard rate function for HCV hazard
}
