% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chzf_ifr_c}
\alias{chzf_ifr_c}
\title{Integrated Hazard rate function - IFR}
\usage{
chzf_ifr_c(tvec, lambda0, thetavec, wvec)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{thetavec}{Locations}

\item{wvec}{Weights}
}
\description{
Integrated hazard rate function for IFR hazard
}
