% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chzf_lcv_c}
\alias{chzf_lcv_c}
\title{Integrated Hazard rate function - LCV}
\usage{
chzf_lcv_c(tvec, lambda0, w0, thetavec, wvec, epsilon)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to scale the hazard rate}

\item{w0}{Offset to be added to the slope of the log hazard rate}

\item{thetavec}{Locations}

\item{wvec}{Weights}

\item{epsilon}{Size of value close to zero}
}
\description{
Integrated hazard rate function for LCV hazard
}
