% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chzf_lwb_c}
\alias{chzf_lwb_c}
\title{Integrated Hazard rate function - LWB}
\usage{
chzf_lwb_c(tvec, lambda0, a, thetavec, wvec)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{a}{Change point}

\item{thetavec}{Locations}

\item{wvec}{Weights}
}
\description{
Integrated hazard rate function for LWB hazard
}
