% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chzf_mew_c}
\alias{chzf_mew_c}
\title{Integrated Hazard rate function - MEW}
\usage{
chzf_mew_c(tvec, alpha, beta, mu, nu)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{alpha}{alpha}

\item{beta}{beta}

\item{mu}{mu}

\item{nu}{nu}
}
\description{
Integrated hazard rate function for MEW hazard
}
