% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chzf_scv_c}
\alias{chzf_scv_c}
\title{Integrated Hazard rate function - SCV}
\usage{
chzf_scv_c(tvec, lambda0, thetavec1, wvec1, thetavec2, wvec2)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{thetavec1}{DFR Locations}

\item{wvec1}{DFR Weights}

\item{thetavec2}{IFR Locations}

\item{wvec2}{IFR Weights}
}
\description{
Integrated hazard rate function for SCV hazard
}
