% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coetzee.R
\docType{data}
\name{coetzee}
\alias{coetzee}
\title{Failure times of rear dump trucks}
\format{
A vector of failure times (in hours)
}
\usage{
data(coetzee)
}
\description{
Data from Coetzee (1996)
}
\examples{
data(coetzee)
\donttest{hist(coetzee, breaks=10, xlab="t (hours)", ylab="Frequency",
               main="Failure times of rear dump trucks")}
}
\references{
Coetzee JL. Reliability degradation and the equipment replacement
problem. In: Proc. Int. Conf. of Maintenance Societies (ICOMS-96).
Melbourne, 1996, Paper 21.
}
\keyword{datasets}
