% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denlife.R
\docType{data}
\name{denlife}
\alias{denlife}
\title{Danish life tables}
\format{
A data frame with five columns: Year, Age, Male, Female, Total
For cohorts of (approximately) 100,000 people born in the years
1835 to 1933, Male/Female/Total give the numbers dying at each age
}
\usage{
data(denlife)
}
\description{
Lifetables for Denmark for cohorts born in 1835 to 1933
}
\examples{
data(denlife)
\donttest{plot(denlife$Age[denlife$Year==1835],
               denlife$Male[denlife$Year==1835],
               main="Age of death for Danish 1835 Male Cohort")}
}
\references{
HMD. Human Mortality Database. Max Planck Institute
for Demographic Research (Germany), University of California,
Berkeley (USA), and French Institute for Demographic Studies (France).
Available at www.mortality.org (data downloaded on 1 August 2025).
}
\keyword{datasets}
