% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{densf}
\alias{densf}
\title{Probability density function}
\usage{
densf(tvec, model.list, use.Cpp = FALSE, log = FALSE)
}
\arguments{
\item{tvec}{Vector of failure times}

\item{model.list}{Model specification}

\item{use.Cpp}{Use C++ functions?  TRUE or FALSE}
}
\description{
Probability density function
}
