% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{drawrgdp}
\alias{drawrgdp}
\title{Plot a draw from the gamma process generated by Stick Breaking (Dirichlet Process)}
\usage{
drawrgdp(alpha = 1, beta = 1, g0 = rexp, ..., kmax = 100)
}
\description{
Plot a draw from the gamma process generated by Stick Breaking (Dirichlet Process)
}
