% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{drawrgrk}
\alias{drawrgrk}
\title{Plot a draw from the gamma process generated by Stick Breaking (RK)}
\usage{
drawrgrk(alpha = 1, beta = 1, gamma = 1, g0 = rexp, ..., kmax = 100)
}
\description{
Plot a draw from the gamma process generated by Stick Breaking (RK)
}
