% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{hazf_chzf}
\alias{hazf_chzf}
\title{Hazard and cumulative hazard function values}
\usage{
hazf_chzf(tvec, model.list, use.Cpp, epsilon = 100 * .Machine$double.neg.eps)
}
\arguments{
\item{tvec}{Vector of failure times}

\item{model.list}{Model specification}

\item{use.Cpp}{Use C++ functions?  TRUE or FALSE}

\item{epsilon}{Small nonzero value}
}
\description{
Hazard and cumulative hazard function values
}
