% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hazf_chzf_hcv_c}
\alias{hazf_chzf_hcv_c}
\title{Hazard and Integrated Hazard rate functions - HCV}
\usage{
hazf_chzf_hcv_c(tvec, lambda0, a, thetavec, wvec)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{a}{Change point}

\item{thetavec}{Locations}

\item{wvec}{Weights}
}
\description{
Hazard rate and integrated hazard rate functions generated by 
a set of discrete locations and weights and integrated
}
