% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hazf_chzf_lcv_c}
\alias{hazf_chzf_lcv_c}
\title{Hazard and Integrated Hazard rate functions - LCV}
\usage{
hazf_chzf_lcv_c(tvec, lambda0, w0, thetavec, wvec, epsilon)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to scale the hazard rate}

\item{w0}{Offset to be added to the slope of the log hazard rate}

\item{thetavec}{Locations}

\item{wvec}{Weights}

\item{epsilon}{Size of value close to zero}
}
\description{
Hazard rate and integrated hazard rate functions generated by 
a set of discrete locations and weights and integrated
}
