% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hazf_chzf_mbt_c}
\alias{hazf_chzf_mbt_c}
\title{Hazard and Integrated Hazard rate functions - MBT}
\usage{
hazf_chzf_mbt_c(tvec, pival, lambda0, thetavec1, wvec1, thetavec2, wvec2)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{pival}{Proportion of the mixture in the DFR component}

\item{lambda0}{Offset to be added to the hazard rate}

\item{thetavec1}{DFR Locations}

\item{wvec1}{DFR Weights}

\item{thetavec2}{IFR Locations}

\item{wvec2}{IFR Weights}
}
\description{
Hazard rate and integrated hazard rate functions generated by 
a set of discrete locations and weights and integrated
}
