% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hazf_chzf_mew_c}
\alias{hazf_chzf_mew_c}
\title{Hazard and Integrated Hazard rate functions - MEW}
\usage{
hazf_chzf_mew_c(tvec, alpha, beta, mu, nu)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{alpha}{alpha}

\item{beta}{beta}

\item{mu}{mu}

\item{nu}{nu}
}
\description{
Hazard rate and integrated hazard rate functions generated by 
a set of discrete locations and weights and integrated
}
