% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hazf_cir_c}
\alias{hazf_cir_c}
\title{Hazard rate function - CIR}
\usage{
hazf_cir_c(tvec, lambda0, thetavec, wvec)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{thetavec}{Locations}

\item{wvec}{Weights}
}
\description{
Hazard rate function for CIR hazard
}
