% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hazf_con_c}
\alias{hazf_con_c}
\title{Hazard rate function - CON}
\usage{
hazf_con_c(tvec, lambda0)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}
}
\description{
Hazard rate function for constant hazard
}
