% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hazf_lcv_c}
\alias{hazf_lcv_c}
\title{Hazard rate function - LCV}
\usage{
hazf_lcv_c(tvec, lambda0, w0, thetavec, wvec)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to scale the hazard rate}

\item{w0}{Offset to be added to the slope of the log hazard rate}

\item{thetavec}{Locations}

\item{wvec}{Weights}
}
\description{
Hazard rate function for LCV hazard
}
