% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hazf_mew_c}
\alias{hazf_mew_c}
\title{Hazard rate function - MEW}
\usage{
hazf_mew_c(tvec, alpha, beta, mu, nu)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{alpha}{alpha}

\item{beta}{beta}

\item{mu}{mu}

\item{nu}{nu}
}
\description{
Hazard rate function for MEW hazard
}
