% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{hist.from.counts}
\alias{hist.from.counts}
\title{Histogram from count data}
\usage{
\method{hist}{from.counts}(counts, mids = NULL, breaks = NULL, col = "light grey", xlim = NULL, ...)
}
\description{
Histogram from count data
}
