% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcfuncs.R
\name{init.objects}
\alias{init.objects}
\title{Initialise objects for an MCMC chain}
\usage{
init.objects(
  tvec,
  n1,
  tau,
  kmax = 100,
  priorpar = list(a1 = 1, a2 = 0.1, b1 = 1, b2 = 0.1, d1 = 1, d2 = 0.1, f1 = 1, f2 =
    0.1, rexp.rate = 1),
  update.par = list(sd.log.alpha = 0.01, sd.logit.v = 0.01, sd.log.theta = 0.01),
  model = "IFR",
  use.Cpp = FALSE,
  seed = NULL,
  generate = "random"
)
}
\description{
Initialise objects for an MCMC chain
}
