% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazfuncs.R
\name{int.lambda.func.cir}
\alias{int.lambda.func.cir}
\title{Integrated hazard rate function - CIR case}
\usage{
int.lambda.func.cir(tvec, lambda0, thetavec, wvec, use.Cpp = TRUE)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}

\item{use.Cpp}{Use cpp to evaluate the function?}
}
\description{
Integrated hazard rate function generated by 
a set of discrete locations and weights and integrated
}
