% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{int.lambda.func.ifr}
\alias{int.lambda.func.ifr}
\title{Integrated hazard rate function - IFR case}
\usage{
int.lambda.func.ifr(tvec, thetavec, wvec, lambda0 = 0, use.Cpp = TRUE)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}

\item{lambda0}{Offset to be added to the hazard rate}

\item{use.Cpp}{Use cpp to evaluate the function?}
}
\description{
Integrated hazard rate function generated by 
a set of discrete locations and weights and integrated
}
