% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{int.lambda.func}
\alias{int.lambda.func}
\title{Integrated hazard rate function}
\usage{
int.lambda.func(tvec, model.list, use.Cpp = TRUE)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{model.list}{Model specification}

\item{use.Cpp}{Use cpp to evaluate the function?}
}
\description{
Hazard rate function generated by 
a set of discrete locations and weights and integrated
}
