% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_lambda_func_ifr_c}
\alias{int_lambda_func_ifr_c}
\title{Integrated hazard rate function - IFR version}
\usage{
int_lambda_func_ifr_c(tvec, thetavec, wvec, lambda0)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}

\item{lambda0}{Offset to be added to the hazard rate (ignored)}
}
\description{
Hazard rate function generated by 
a set of discrete locations and weights and integrated
}
