% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_lambda_func_lcv_c}
\alias{int_lambda_func_lcv_c}
\title{Integrated hazard rate function - Log Convex version}
\usage{
int_lambda_func_lcv_c(tvec, lambda0, w0, thetavec, wvec, epsilon)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{hazard rate scaling}

\item{w0}{hazard rate gradient offset}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}
}
\description{
Integrated hazard rate function generated by 
a set of discrete locations and weights and integrated
}
