% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int_lambda_func_lwb_c}
\alias{int_lambda_func_lwb_c}
\title{Integrated hazard rate function - LWB version}
\usage{
int_lambda_func_lwb_c(tvec, lambda0, a, thetavec, wvec)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate (ignored)}

\item{a}{inflection point}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}
}
\description{
Hazard rate function generated by 
a set of discrete locations and weights and integrated
}
