% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlefuncs.R
\name{inthaz.gr.mew}
\alias{inthaz.gr.mew}
\title{Gradient of integrated hazard function the MEW function}
\usage{
inthaz.gr.mew(tvec, lambda, alpha, theta, gamma)
}
\description{
Parameterisation is lambda,alpha,theta,gamma
}
