% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{invsurvf}
\alias{invsurvf}
\title{Inverse Survival function}
\usage{
invsurvf(uvec, model.list, use.Cpp = FALSE)
}
\arguments{
\item{uvec}{Vector of survival function values}

\item{model.list}{Model specification}

\item{use.Cpp}{Use C++ functions?  TRUE or FALSE}
}
\description{
Inverse Survival function
}
