% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invsurvf_con_c}
\alias{invsurvf_con_c}
\title{Inverse survival function - CON}
\usage{
invsurvf_con_c(uvec, lambda0)
}
\arguments{
\item{uvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}
}
\description{
Inverse survival function for constant hazard
}
