% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{lambda.func.ifr}
\alias{lambda.func.ifr}
\title{Hazard rate function - IFR case}
\usage{
lambda.func.ifr(tvec, thetavec, wvec, lambda0 = 0, use.Cpp = TRUE)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}

\item{lambda0}{Offset to be added to the hazard rate}

\item{use.Cpp}{Use C++ function?}
}
\description{
Hazard rate function - IFR case
}
