% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda_func_cvx_c}
\alias{lambda_func_cvx_c}
\title{Hazard rate function - CVX}
\usage{
lambda_func_cvx_c(tvec, lambda0, tau, thetavec1, wvec1, thetavec2, wvec2)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}
}
\description{
Hazard rate function generated by 
a set of discrete locations and weights and integrated
}
