% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda_func_lwb_c}
\alias{lambda_func_lwb_c}
\title{Hazard rate function - LWB}
\usage{
lambda_func_lwb_c(tvec, lambda0, a, thetavec, wvec)
}
\arguments{
\item{tvec}{Locations at which to evaluate the function}

\item{lambda0}{Offset to be added to the hazard rate}

\item{a}{inflection point}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}
}
\description{
Hazard rate function - LWB
}
