% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{llfunc}
\alias{llfunc}
\title{Log likelihood function}
\usage{
llfunc(datlist, model.list, verbose = FALSE)
}
\arguments{
\item{verbose}{Display output in iterative evaluations?  TRUE or FALSE}

\item{tvec}{Vector of failure times}

\item{parvec}{Parameter vector}

\item{model}{Model code (three letters)}
}
\description{
Log likelihood function
}
