% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazfuncs.R
\name{llike.fail}
\alias{llike.fail}
\title{Log likelihood of failure data}
\usage{
llike.fail(tvec, model.list, tau = Inf, use.Cpp = TRUE)
}
\arguments{
\item{tvec}{Vector of failure times}

\item{model.list}{Model specification}

\item{tau}{Censoring time}

\item{use.Cpp}{Use C++ functions for speed}
}
\description{
Log likelihood of failure data
}
