% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logprior_ifrdfr_c}
\alias{logprior_ifrdfr_c}
\title{Log prior - IFR/DFR}
\usage{
logprior_ifrdfr_c(
  eta,
  gamma,
  thetavec,
  vvec,
  alpha,
  beta,
  phi,
  nu,
  a1,
  a2,
  b1,
  b2,
  f1,
  f2
)
}
\arguments{
\item{eta}{eta}

\item{gamma}{gamma}
}
\description{
log(prior) for the IFR and DFR models - in vector form
}
