% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logprior_lcv_c}
\alias{logprior_lcv_c}
\title{Log prior - LCV}
\usage{
logprior_lcv_c(
  lambda0,
  w0,
  gamma,
  thetavec,
  vvec,
  a,
  alpha,
  beta,
  nu,
  phi,
  s1,
  s2,
  sigmapw0,
  a1,
  a2,
  b1,
  b2,
  g1,
  g2,
  f1,
  f2
)
}
\arguments{
\item{gamma}{gamma}

\item{eta}{eta}
}
\description{
log(prior) for the LCV model - in vector form
}
