% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logprior_mew_c}
\alias{logprior_mew_c}
\title{Log prior - MEW}
\usage{
logprior_mew_c(alpha, beta, mu, nu, a1, a2, b1, b2, s1, s2, t1, t2)
}
\arguments{
\item{alpha}{alpha}

\item{beta}{beta}

\item{mu}{mu}

\item{nu}{nu}

\item{a1}{a1 Prior for alpha}

\item{a2}{a2 Prior for alpha}

\item{b1}{b1 Prior for beta}

\item{b2}{b2 Prior for beta}

\item{s1}{s1 Prior for mu}

\item{s2}{s2 Prior for mu}

\item{t1}{t1 Prior for nu}

\item{t2}{t2 Prior for nu}
}
\description{
log(prior) for the MEW model - in vector form
}
