% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lprior_circdr_c}
\alias{lprior_circdr_c}
\title{Log prior - CIR/CDR}
\usage{
lprior_circdr_c(
  eta,
  gamma,
  thetavec,
  vvec,
  alpha,
  beta,
  phi,
  nu,
  a1,
  a2,
  b1,
  b2,
  f1,
  f2
)
}
\arguments{
\item{eta}{eta}

\item{gamma}{gamma}
}
\description{
log(prior) for the CIR and CDR models - in vector form
}
