% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lprior_lwb_c}
\alias{lprior_lwb_c}
\title{Log prior - LWB}
\usage{
lprior_lwb_c(
  eta,
  a,
  gamma,
  thetavec,
  vvec,
  alpha,
  beta,
  phi,
  nu,
  c1,
  c2,
  a1,
  a2,
  b1,
  b2,
  f1,
  f2
)
}
\arguments{
\item{eta}{eta}

\item{gamma}{gamma}
}
\description{
log(prior) for the LWB model - in vector form
}
