% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lprior_mbt_c}
\alias{lprior_mbt_c}
\title{Log prior - MBT}
\usage{
lprior_mbt_c(
  eta1,
  gamma1,
  thetavec1,
  vvec1,
  alpha1,
  beta1,
  phi1,
  eta2,
  gamma2,
  thetavec2,
  vvec2,
  alpha2,
  beta2,
  phi2,
  nu,
  a1,
  a2,
  b1,
  b2,
  f11,
  f21,
  f12,
  f22
)
}
\arguments{
\item{eta}{eta}

\item{gamma}{gamma}
}
\description{
log(prior) for the MBT model - in vector form
}
