% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lprior_mew_c}
\alias{lprior_mew_c}
\title{Log prior - MEW}
\usage{
lprior_mew_c(lambda, alpha, theta, gamma, s1, s2, a1, a2, t1, t2, g1, g2)
}
\arguments{
\item{lambda}{lambda}

\item{alpha}{alpha}

\item{theta}{theta}

\item{gamma}{gamma}
}
\description{
log(prior) for the MEW model - in vector form
}
