% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-modelfuncs.R
\name{lpriorf.vector.old}
\alias{lpriorf.vector.old}
\title{Components of the log prior of the current state - return as a vector}
\usage{
lpriorf.vector.old(state, fpar, model)
}
\description{
Components of the log prior of the current state - return as a vector
}
