% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs.R
\name{lpriorf.vector}
\alias{lpriorf.vector}
\title{Components of the log prior of the current state - return as a vector}
\usage{
lpriorf.vector(state, fpar, model)
}
\description{
Components of the log prior of the current state - return as a vector
}
