% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{lpriorfunc.vector}
\alias{lpriorfunc.vector}
\title{Log prior function - vector valued}
\usage{
lpriorfunc.vector(state, fpar, use.Cpp = TRUE)
}
\arguments{
\item{use.Cpp}{Use C++ functions?}

\item{model.list}{Model specification}
}
\description{
Log prior function - vector valued
}
