% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{make.datlist}
\alias{make.datlist}
\title{Make data list object}
\usage{
make.datlist(tvec, obs = NULL)
}
\arguments{
\item{tvec}{Vector of failure times}

\item{obs}{Vector of TRUE (observed) or FALSE (right censored) indicators}
}
\description{
Make data list object
}
