% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatefuncs.R
\name{make.state}
\alias{make.state}
\title{Construct the state object from the parameter vectors and data}
\usage{
make.state(epar, datlist, fpar, ppar, model)
}
\description{
Construct the state object from the parameter vectors and data
}
