% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makev_c}
\alias{makev_c}
\title{Make DPP beta draws given a set of unscaled stick breaking weights}
\usage{
makev_c(uvec)
}
\description{
Make DPP beta draws given a set of unscaled stick breaking weights
}
