% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{makew}
\alias{makew}
\title{Make the DPP Stick Breaking Weights}
\usage{
makew(vvec)
}
\arguments{
\item{vvec}{Vector of draws from Beta(1,alpha)}
}
\description{
Make the DPP stick breaking weights from a set of 
draws from Beta(1,alpha)
}
