% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeker.R
\docType{data}
\name{meeker}
\alias{meeker}
\title{Failure times of 30 devices}
\format{
A vector of failure times (in hours)
}
\usage{
data(meeker)
}
\description{
Data from Meeker et al. (2022)
}
\examples{
data(meeker)
\donttest{hist(coetzee, breaks=15, xlab="t (hours)", ylab="Frequency",
               main="Failure times of 30 devices from Meeker et al. (2022)")}
}
\references{
W. Q. Meeker, L. A. Escobar, and F. G. Pascual, 
            Statistical Methods for Reliability Data. 
            Hoboken, NJ, USA: Wiley, 2022.
}
\keyword{datasets}
