% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{mlfit}
\alias{mlfit}
\title{Maximum likelihood fitting for a restricted set of models}
\usage{
mlfit(model, tvec, obs = TRUE)
}
\arguments{
\item{model}{Chosen model}

\item{tvec}{Vector of failure times}

\item{obs}{Vector of logical values indicating observation or right-censoring}
}
\description{
Maximum likelihood fitting for a restricted set of models
}
