% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{ncompf}
\alias{ncompf}
\title{Count the number of components adding to fixed proportion}
\usage{
ncompf(state, proportion = 0.95, datlist, fpar, ppar, model)
}
\description{
Count the number of components adding to fixed proportion
}
