% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZDeaths2023.R
\docType{data}
\name{NZDeaths2023}
\alias{NZDeaths2023}
\title{Deaths in New Zealand in 2023}
\format{
Counts of deaths in New Zeland in 2023 by age and sex.
}
\usage{
data(NZDeaths2023)
}
\description{
Data from Statistics New Zealand
}
\examples{
data(NZDeaths2023)
\donttest{barplot(t(as.matrix(NZDeaths2023[,c("Male","Female")])),
   beside=TRUE,names=dcounts$age_group, las=2, cex.lab=1.0, 
   cex.names=0.8, cex.axis=0.8,
   legend=TRUE, args.legend=list(x="topleft"),
   xlab="Age group", ylab="Counts", main="Deaths in 2023")
          mtext("Source: Statistics New Zealand", 
                side=1, line=3, adj=1, cex=0.5)}
}
\references{
Statistics New Zealand \url{https://www.stats.govt.nz/topics/population}
}
\keyword{datasets}
