% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlefuncs.R
\name{ofunc.gr.mew}
\alias{ofunc.gr.mew}
\title{Gradient of Log Likelihood of the MEW function for use by optim()}
\usage{
ofunc.gr.mew(parvec, datlist, fpar, ppar, verbose = FALSE)
}
\description{
Parameterisation is log(alpha),log(beta),log(mu),log(nu)
}
