% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlefuncs.R
\name{ofunc.mew}
\alias{ofunc.mew}
\title{Log Likelihood of the MEW function for use by optim()}
\usage{
ofunc.mew(parvec, datlist, fpar, ppar, verbose = FALSE)
}
\description{
Parameterisation is log(lambda),log(alpha),log(theta),log(gamma)
}
