% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{plot.hazf}
\alias{plot.hazf}
\title{Plot a hazard function}
\usage{
\method{plot}{hazf}(
  model.list,
  xlim,
  use.Cpp = FALSE,
  n = 101,
  add = FALSE,
  xlab = "t",
  ylab = expression(lambda(t)),
  scale = 1,
  ...
)
}
\description{
Plot a hazard function
}
