% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{plot_chzf}
\alias{plot_chzf}
\title{Plot a cumulative hazard function}
\usage{
plot_chzf(
  model.list,
  xlim,
  use.Cpp = FALSE,
  n = 101,
  add = FALSE,
  xlab = "t",
  ylab = expression(Lambda(t)),
  scale = 1,
  ...
)
}
\description{
Plot a cumulative hazard rate function
}
