% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{plot_densf}
\alias{plot_densf}
\title{Plot a density function}
\usage{
plot_densf(
  model.list,
  xlim,
  use.Cpp = FALSE,
  n = 101,
  add = FALSE,
  xlab = "t",
  ylab = expression(f(t)),
  scale = 1,
  ...
)
}
\description{
Plot a density function
}
