% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazfuncs.R
\name{plot_densityfunc}
\alias{plot_densityfunc}
\title{Plot probability density function}
\usage{
plot_densityfunc(model.list, npts = 101, main = NULL, use.Cpp = TRUE, ...)
}
\arguments{
\item{model.list}{Model specification}

\item{npts}{Number of plotting points}

\item{use.Cpp}{Use C++ code?}

\item{...}{Other arguments to plot()}
}
\description{
Plot probability density function
}
