% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{plot_invsurvf}
\alias{plot_invsurvf}
\title{Plot an inverse survival function}
\usage{
plot_invsurvf(
  model.list,
  xlim = c(0, 1) + 1e-04 * c(1, -1),
  use.Cpp = FALSE,
  n = 101,
  add = FALSE,
  xlab = expression(bar(F)(t)),
  ylab = "t",
  scale = 1,
  ...
)
}
\description{
Plot an inverse survival function
}
