% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazfuncs.R
\name{plot_lambdafunc}
\alias{plot_lambdafunc}
\title{Plot hazard rate function}
\usage{
plot_lambdafunc(
  model.list,
  npts = 101,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  add = FALSE,
  use.Cpp = TRUE,
  ...
)
}
\arguments{
\item{model.list}{Model specification}

\item{npts}{Number of plotting points}

\item{use.Cpp}{Use C++ code?}

\item{...}{Other arguments to plot()}
}
\description{
Plot hazard rate function
}
