% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs.R
\name{plot_state}
\alias{plot_state}
\title{Plot a representation of the current state}
\usage{
plot_state(
  state,
  datlist,
  fpar,
  ppar,
  model,
  type = 1,
  main = NULL,
  add = FALSE,
  legend.loc = "topright",
  ...
)
}
\arguments{
\item{state}{The current state}

\item{datlist}{The data list}

\item{fpar}{The fixed parameters}

\item{ppar}{Proposal parameters}

\item{model}{Type of model (IFR or DFR)}

\item{type}{Type of plot (ignored at the moment)}

\item{add}{Logical: add the state to an existing plot?}
}
\description{
Plot the current state
}
