% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{plot_survf}
\alias{plot_survf}
\title{Plot a survival function}
\usage{
plot_survf(
  model.list,
  xlim,
  ylim = c(0, 1),
  use.Cpp = FALSE,
  n = 101,
  add = FALSE,
  xlab = "t",
  ylab = expression(bar(F)(t)),
  scale = 1,
  ...
)
}
\description{
Plot a survival function
}
