% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazfuncs.R
\name{plot_survivalfunc}
\alias{plot_survivalfunc}
\title{Plot survival function}
\usage{
plot_survivalfunc(model.list, npts = 101, use.Cpp = TRUE, ...)
}
\arguments{
\item{model.list}{Model specification}

\item{npts}{Number of plotting points}

\item{use.Cpp}{Use C++ code?}

\item{...}{Other arguments to plot()}
}
\description{
Plot survival function
}
