% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazfuncs.R
\name{rfail.cir}
\alias{rfail.cir}
\title{Simulate from a CIR model}
\usage{
rfail.cir(n, lambda0, thetavec, wvec)
}
\arguments{
\item{n}{Number of failure times to simulate}

\item{lambda0}{Offset to be added to the hazard rate}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}
}
\description{
Simulate from a CIR model
}
