% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazfuncs.R
\name{rfail.dfr}
\alias{rfail.dfr}
\title{Simulate from an DFR model}
\usage{
rfail.dfr(n, lambda0, thetavec, wvec)
}
\arguments{
\item{n}{Number of failure times to simulate}

\item{lambda0}{Offset to be added to the hazard rate}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}
}
\description{
Simulate from an DFR model
}
