% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{rfail}
\alias{rfail}
\title{Simulate from a hazard rate function}
\usage{
rfail(n, model.list)
}
\arguments{
\item{n}{Number of failure times to simulate}

\item{model.list}{Model specification list}

\item{model}{IFR or DFR (increasing or decreasing failure rates)}
}
\description{
Simulate from a hazard rate function generated by 
a set of discrete locations and weights and integrated
}
