% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{rgammp}
\alias{rgammp}
\title{Generate from a Gamma Process}
\usage{
rgammp(tvec, alpha = 1, beta = 1)
}
\arguments{
\item{tvec}{Vector of locations at which to simulate}

\item{alpha}{Scaling of the alpha() measure [constant]}

\item{beta}{Weighting scale}
}
\description{
Generate from a Gamma Process
}
