% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{rgammp_bridge}
\alias{rgammp_bridge}
\title{Generate from a Gamma Bridge}
\usage{
rgammp_bridge(tvec, y0 = 0, y1 = NULL, alpha = 1, beta = 1)
}
\arguments{
\item{tvec}{Vector of locations at which to simulate}

\item{y0}{Value of the process at first tvec}

\item{y1}{Value of the process at last tvec (unconstrained if NULL)}

\item{alpha}{Scaling of the alpha() measure [constant]}

\item{beta}{Weighting scale}
}
\description{
Generate from a Gamma Bridge
}
