% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{rgammp_sbrk}
\alias{rgammp_sbrk}
\title{Generate from a Gamma process using the RK Stick breaking construction}
\usage{
rgammp_sbrk(alpha = 1, beta = 1, gamma = 1, g0 = rexp, ..., kmax = 100)
}
\arguments{
\item{alpha}{Scaling of the alpha() measure [constant]}

\item{beta}{Weighting scale}

\item{g0}{Base measure (see Details)}

\item{...}{Arguments to g0}

\item{kmax}{Truncation value}
}
\description{
Generate from a Gamma process using the RK Stick breaking construction
}
\details{
\code{g0(n, ...)} is a function that generates \code{n} values from a random
distribution with parameters \code{...} .   For example \code{g0=rexp} (the default) generates
from an Exponential distribution.
}
