% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{rgaussp}
\alias{rgaussp}
\title{Generate from a Gaussian process}
\usage{
rgaussp(tvec, sigma = 1)
}
\arguments{
\item{tvec}{Vector of locations}

\item{sigma}{Standard deviation}
}
\description{
Generate from a Gaussian process
}
