% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{rgaussp_bridge}
\alias{rgaussp_bridge}
\title{Generate from a Gaussian Bridge}
\usage{
rgaussp_bridge(tvec, y0 = 0, y1 = NULL, sigma = 1)
}
\arguments{
\item{tvec}{Vector of locations at which to simulate}

\item{y0}{Value of the process at first tvec}

\item{y1}{Value of the process at last tvec (unconstrained if NULL)}

\item{sigma}{Standard deviation}
}
\description{
Generate from a Gaussian Bridge
}
