% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{rgen}
\alias{rgen}
\title{Generate failure times from a model}
\usage{
rgen(n, model.list, use.Cpp = FALSE, seed = NULL)
}
\arguments{
\item{n}{Number of failure times to simulate}

\item{model.list}{Model specification}

\item{use.Cpp}{Use C++ functions?  TRUE or FALSE}

\item{seed}{Random number seed}
}
\description{
Generate failure times from a model
}
