% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcfuncs.R
\name{run.chain}
\alias{run.chain}
\title{Run an MCMC chain}
\usage{
run.chain(
  nstore,
  nthin = 1,
  nburn = 0,
  state,
  datlist,
  fpar,
  ppar,
  model,
  do.plot = FALSE,
  show.progress = NULL,
  tryupdate = FALSE
)
}
\description{
Run an MCMC chain
}
