% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcfuncs.R
\name{run.multiple.chains}
\alias{run.multiple.chains}
\title{Run multiple MCMC chains - with starting states specified in state.list}
\usage{
run.multiple.chains(
  state.list,
  nstore,
  nthin = 1,
  nburn = 0,
  datlist,
  fpar,
  ppar,
  model,
  do.plot = FALSE,
  show.progress = NULL
)
}
\description{
Run multiple MCMC chains - with starting states specified in state.list
}
