% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{sample.datlist}
\alias{sample.datlist}
\title{Subsample within data list}
\usage{
sample.datlist(n, datlist)
}
\arguments{
\item{n}{The number of elements to sample}

\item{datlist}{A datlist object}
}
\description{
Subsample within data list
}
