% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcfuncs.R
\name{state.as.vector}
\alias{state.as.vector}
\title{Convert the state from a list to a vector representation}
\usage{
state.as.vector(state, fpar, model)
}
\description{
Convert the state from a list to a vector representation
}
