% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.a.v1}
\alias{update.a.v1}
\title{a V1 (LWB)}
\usage{
\method{update}{a.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(lambda0 = "lambda0", gamma = "gamma", thetavec = "thetavec", vvec = "vvec",
    alpha = "alpha", beta = "beta", nu = "nu", phi = "phi", uvec = "uvec", wvec = "wvec")
)
}
\description{
a V1 (LWB)
}
