% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.alpha.mew}
\alias{update.alpha.mew}
\title{alpha (MH) MEW}
\usage{
\method{update}{alpha.mew}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(lambda = "lambda", alpha = "alpha", theta = "theta", gamma = "gamma")
)
}
\description{
alpha (MH) MEW
}
