% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.eta.v1}
\alias{update.eta.v1}
\title{eta V1 (IFR/DFR/LWB/SBT/MBT)
(Metropolis-Hastings update: log Normal proposal)}
\usage{
\method{update}{eta.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(eta = "eta", gamma = "gamma", thetavec = "thetavec", vvec = "vvec", alpha =
    "alpha", beta = "beta", phi = "phi", uvec = "uvec", wvec = "wvec", lambda0 =
    "lambda0", thetaswap = "thetaswap")
)
}
\description{
eta V1 (IFR/DFR/LWB/SBT/MBT)
(Metropolis-Hastings update: log Normal proposal)
}
