% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.gamma.v1}
\alias{update.gamma.v1}
\title{gamma V1 (Gibbs) (IFR/DFR/LWB)}
\usage{
\method{update}{gamma.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(eta = "eta", gamma = "gamma", thetavec = "thetavec", vvec = "vvec", alpha =
    "alpha", beta = "beta", phi = "phi", uvec = "uvec", wvec = "wvec", lambda0 =
    "lambda0", thetaswap = "thetaswap")
)
}
\description{
gamma V1 (Gibbs) (IFR/DFR/LWB)
}
