% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.gamma.v2}
\alias{update.gamma.v2}
\title{gamma V2 (MH) (SBT,MBT,LCV)}
\usage{
\method{update}{gamma.v2}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(lambda0 = "lambda0", w0 = "w0", gamma = "gamma", thetavec = "thetavec", vvec =
    "vvec", alpha = "alpha", beta = "beta", phi = "phi", uvec = "uvec", wvec = "wvec",
    thetaswap = "thetaswap")
)
}
\description{
gamma V2 (MH) (SBT,MBT,LCV)
}
