% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.lambda0.v0}
\alias{update.lambda0.v0}
\title{lambda0 V0 (Gibbs) (CON)}
\usage{
\method{update}{lambda0.v0}(state, datlist, fpar, ppar, model, nm = c(lambda0 = "lambda0"))
}
\description{
lambda0 V0 (Gibbs) (CON)
}
