% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.lambda0.v2}
\alias{update.lambda0.v2}
\title{lambda0 V2 (Gibbs) (LCV)}
\usage{
\method{update}{lambda0.v2}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(lambda0 = "lambda0", w0 = "w0", gamma = "gamma", thetavec = "thetavec", vvec =
    "vvec", alpha = "alpha", beta = "beta", nu = "nu", phi = "phi", uvec = "uvec", wvec =
    "wvec")
)
}
\description{
lambda0 V2 (Gibbs) (LCV)
}
