% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.mu.mew}
\alias{update.mu.mew}
\title{mu (MH) MEW}
\usage{
\method{update}{mu.mew}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(alpha = "alpha", beta = "beta", mu = "mu", nu = "nu")
)
}
\description{
mu (MH) MEW
}
