% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.phi.v1}
\alias{update.phi.v1}
\title{phi V1 (IFR/DFR)}
\usage{
\method{update}{phi.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(lambda0 = "lambda0", gamma = "gamma", thetavec = "thetavec", vvec = "vvec",
    alpha = "alpha", beta = "beta", nu = "nu", phi = "phi", uvec = "uvec", wvec = "wvec",
    g1 = "g1", g2 = "g2", f1 = "f1", f2 = "f2")
)
}
\description{
phi V1 (IFR/DFR)
}
