% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.thetavec.v1}
\alias{update.thetavec.v1}
\title{thetavec V1 (IFR/DFR)}
\usage{
\method{update}{thetavec.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(lambda0 = "lambda0", gamma = "gamma", thetavec = "thetavec", vvec = "vvec",
    alpha = "alpha", beta = "beta", nu = "nu", phi = "phi", uvec = "uvec", wvec = "wvec")
)
}
\description{
thetavec V1 (IFR/DFR)
}
