% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.vvec.v1}
\alias{update.vvec.v1}
\title{vvec V1 (IFR/DFR)}
\usage{
\method{update}{vvec.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(eta = "eta", gamma = "gamma", thetavec = "thetavec", vvec = "vvec", alpha =
    "alpha", beta = "beta", phi = "phi", uvec = "uvec", wvec = "wvec", lambda0 =
    "lambda0", thetaswap = "thetaswap")
)
}
\description{
vvec V1 (IFR/DFR)
}
