% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.w0.v1}
\alias{update.w0.v1}
\title{w0 V1 (LCV)}
\usage{
\method{update}{w0.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(lambda0 = "lambda0", w0 = "w0", gamma = "gamma", thetavec = "thetavec", vvec =
    "vvec", alpha = "alpha", beta = "beta", phi = "phi", uvec = "uvec", wvec = "wvec",
    thetaswap = "thetaswap")
)
}
\description{
w0 V1 (LCV)
}
