% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.wvec.v1}
\alias{update.wvec.v1}
\title{wvec V1 (IFR/DFR/LWB/SBT/MBT/LCV)}
\usage{
\method{update}{wvec.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(lambda0 = "lambda0", gamma = "gamma", thetavec = "thetavec", vvec = "vvec",
    alpha = "alpha", beta = "beta", nu = "nu", phi = "phi", uvec = "uvec", wvec = "wvec",
    lambda0 = "lambda0")
)
}
\description{
wvec V1 (IFR/DFR/LWB/SBT/MBT/LCV)
}
