% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcfuncs.R
\name{vector.as.state}
\alias{vector.as.state}
\title{Convert the state from a vector to a list representation}
\usage{
vector.as.state(statevec, datlist, fpar, ppar, model)
}
\description{
Convert the state from a vector to a list representation
}
