% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{as.model.list}
\alias{as.model.list}
\title{Convert parameter vector to model list object (incomplete)}
\usage{
as.model.list(parvec, model)
}
\arguments{
\item{model.list}{Model specification}
}
\description{
Suitable for calls to optim, and only for models CON and MEW
}
