% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatefuncs.R
\name{init.objects}
\alias{init.objects}
\title{Initialise objects for an MCMC chain}
\usage{
init.objects(
  tvec,
  obs,
  kmax = 100,
  prior.par = NULL,
  update.par = NULL,
  model = "IFR",
  use.Cpp = FALSE,
  seed = NULL,
  generate = "random"
)
}
\description{
Initialise objects for an MCMC chain
}
