% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logprior_con_c}
\alias{logprior_con_c}
\title{Log prior - CON}
\usage{
logprior_con_c(lambda0, s1, s2)
}
\arguments{
\item{lambda0}{lambda0: constant hazard rate}

\item{nu}{exponential prior}
}
\description{
log(prior) for the CON model - in vector form
}
