% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{rfail.ifr}
\alias{rfail.ifr}
\title{Simulate from an IFR model}
\usage{
rfail.ifr(n, thetavec, wvec, lambda0 = 0)
}
\arguments{
\item{n}{Number of failure times to simulate}

\item{thetavec}{Set of support locations}

\item{wvec}{Set of associated weights}

\item{lambda0}{Offset to be added to the hazard rate}
}
\description{
Simulate from an IFR model
}
