% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazfuncs.R
\name{rfail}
\alias{rfail}
\title{Simulate from a hazard rate function}
\usage{
rfail(n, model.list, tau = Inf)
}
\arguments{
\item{n}{Number of failure times to simulate}

\item{model.list}{Model specification list}

\item{model}{Label of the model - see Details}
}
\description{
Simulate from a hazard rate function generated by 
a set of discrete locations and weights and integrated

The \code{model} label can be one of 
IFR (increasing failure rate),
DFR (decreasing failure rate),
LWB (Lo Weng Bathtub),
SBT (Superposition Bathtub),
MBT (Mixture Bathtub), 
LCV (Log Convex Bathtub)
MEW (Modified Exponentiated Weibull)
}
