% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfuncs2.R
\name{survf}
\alias{survf}
\title{Survival function}
\usage{
survf(tvec, model.list, use.Cpp = FALSE)
}
\arguments{
\item{tvec}{Vector of failure times}

\item{model.list}{Model specification}

\item{use.Cpp}{Use C++ functions?  TRUE or FALSE}
}
\description{
Survival function
}
