% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.a.v1}
\alias{update.a.v1}
\title{a V1 (LWB)}
\usage{
\method{update}{a.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(eta = "eta", gamma = "gamma", thetavec = "thetavec", vvec = "vvec", alpha =
    "alpha", beta = "beta", phi = "phi", uvec = "uvec", wvec = "wvec", lambda0 =
    "lambda0", thetaswap = "thetaswap")
)
}
\description{
a V1 (LWB)
}
