% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-subfuncs.R
\name{update.beta.v1}
\alias{update.beta.v1}
\title{beta V1 (IFR/DFR/LWB/SBT/MBT/LCV)}
\usage{
\method{update}{beta.v1}(
  state,
  datlist,
  fpar,
  ppar,
  model,
  nm = c(eta = "eta", gamma = "gamma", thetavec = "thetavec", vvec = "vvec", alpha =
    "alpha", beta = "beta", phi = "phi", uvec = "uvec", wvec = "wvec", lambda0 =
    "lambda0", thetaswap = "thetaswap")
)
}
\description{
beta V1 (IFR/DFR/LWB/SBT/MBT/LCV)
}
